feat: add jump component with gravity inside
